AI architecture/requirements.txt
langchain==0.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.0.1
streamlit==1.31.1 